updated mylaser.py
I used the pdb library to debug a lot of the problems such as incorrect function syntax and the scan ranges data being stored in tuples. I had to convert the tuples to a list using the list() command. I've noticed that the LaserScan gives an error message when the average value of the 10 points that we have selected is below 0.47. I do not know why. We will have to ask Dr. Vela regarding his input regarding this problem.